split check & patch
